requirements.txt
pycryptodome
numpy
Pillow